Move tests to live next to source files
