+++ app/src/utils/add-related-primary-key-to-fields.test.ts
@@ -2,7 +2,7 @@ import { test, expect, vi, beforeEach, afterEach, Mock } from 'vitest';
 import { setActivePinia } from 'pinia';
 import { createTestingPinia } from '@pinia/testing';
 
-import { cryptoStub } from '../__stubs__/crypto';
+import { cryptoStub } from '../__test_utils__/crypto';
 vi.stubGlobal('crypto', cryptoStub);
 
 import { useFieldsStore } from '@/stores/fields';
