+++ app/src/utils/capitalize-first.test.ts
