+++ app/src/utils/flatten-field-groups.test.ts
